[kernel] define extra tag to distinguish patch versions
Since we are starting to support download pre-built kernel packages,
every time we add any change to the kernel. We have to make some
distinctive mark in the kernel version so that we know what version
we are building/downloading.

This is the infrastructure work to get kernel subversion updated in
target.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>